rebase and fix for broken 3.10
