+++ tox.ini
@@ -269,6 +269,7 @@ commands =
 
 [testenv:local]
 usedevelop = true
+basepython = python3.9
 deps =
     {[base]deps}
     {[dev]deps}
